jergen-credit-ai/Jergen AI/requirements.txt
anthropic>=0.39.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.5.0
python-docx>=1.1.0
docx2pdf>=0.1.8
jellyfish>=0.9.0
